Move Camera into camera.rs
